requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astor==0.8.1
astropy==4.1
beautifulsoup4==4.6.0
cached-property==1.5.2
certifi==2020.12.5
cftime==1.3.0
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decode==0.5.9
Django==3.1.4
djangorestframework==3.12.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
joblib==1.0.0
JPype1==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
konlpy==0.5.2
lxml==4.6.2
Markdown==3.3.3
matplotlib==3.3.3
mysqlclient==2.0.2
netCDF4==1.5.5.1
numpy==1.19.3
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.0.1
protobuf==3.14.0
PyMySQL==0.10.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.22.2.post1
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
tensorboard==1.15.0
tensorflow==1.15.3
tensorflow-estimator==1.15.1
termcolor==1.1.0
tqdm==4.54.1
tweepy==3.9.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
xarray==0.15.1
zipp==3.4.0
